refactor(config): migrate to map-based TOML parsing and add gap management
- Replace regex-based edits with map-based updates; introduce MonitorGap, Summary, and updateMonitorGapInConfig.
- Switch to BurntSushi/toml encoder for writes and restructure AerospaceService to hold parsed config.
- Update per-monitor gap keys to gaps.outer.left.monitor and gaps.outer.right.monitor; tests adjusted accordingly.
- Bump dependencies to align with new TOML handling (toml, cli, x/sys).
- BREAKING CHANGE: gaps config layout changed; consumers must adapt to the new TOML structure and the write behavior may reformat the config.

+++ cmd/workspace/adjust.go
@@ -61,7 +61,7 @@ func runAdjust(cmd *ufcli.Command) error {
 	}
 
 	// Calculate new percentage
-	newPercent := *monState.Current + amount
+	newPercent := *monState.Current + int64(amount)
 
 	// Validate new percentage
 	if err := gaps.ValidatePercentage(newPercent); err != nil {
